feat(ESSNTL-4253): Filtering refactor (#1297)
* Refactor & cleanup

* Refactor system profile
