+++ api/filtering/filtering.py
@@ -15,6 +15,8 @@ from api.filtering.filtering_common import SUPPORTED_FORMATS
 from app import custom_filter_fields
 from app import inventory_config
 from app import system_profile_spec
+from app.auth import get_current_identity
+from app.auth.identity import IdentityType
 from app.exceptions import ValidationException
 from app.logging import get_logger
 from app.utils import Tag
@@ -331,8 +333,12 @@ def build_tag_query_dict_tuple(tags):
     return query_tag_tuple
 
 
+def owner_id_filter():
+    return ({"spf_owner_id": {"eq": get_current_identity().system["cn"]}},)
+
+
 def host_id_list_query_filter(host_id_list):
-    return (
+    all_filters = (
         {
             "stale_timestamp": {
                 "gt": str((datetime.now(timezone.utc) - inventory_config().culling_culled_offset_delta).isoformat())
@@ -346,20 +352,26 @@ def host_id_list_query_filter(host_id_list):
         },
     )
 
+    current_identity = get_current_identity()
+    if current_identity.identity_type == IdentityType.SYSTEM:
+        all_filters += owner_id_filter()
+
+    return all_filters
+
 
 def query_filters(
-    fqdn,
-    display_name,
-    hostname_or_id,
-    insights_id,
-    provider_id,
-    provider_type,
-    updated_start,
-    updated_end,
-    tags,
-    staleness,
-    registered_with,
-    filter,
+    fqdn=None,
+    display_name=None,
+    hostname_or_id=None,
+    insights_id=None,
+    provider_id=None,
+    provider_type=None,
+    updated_start=None,
+    updated_end=None,
+    tags=None,
+    staleness=None,
+    registered_with=None,
+    filter=None,
 ):
     num_ids = 0
     for id_param in [fqdn, display_name, hostname_or_id, insights_id]:
@@ -413,6 +425,10 @@ def query_filters(
             query_filters += build_system_profile_filter(filter["system_profile"])
         else:
             raise ValidationException("filter key is invalid")
+
+    current_identity = get_current_identity()
+    if current_identity.identity_type == IdentityType.SYSTEM:
+        query_filters += owner_id_filter()
 
     logger.debug(query_filters)
     return query_filters
