+++ api/system_profile.py
@@ -8,16 +8,11 @@ from api import build_collection_response
 from api import custom_escape
 from api import flask_json_response
 from api import metrics
-from api.filtering.filtering import build_registered_with_filter
-from api.filtering.filtering import build_system_profile_filter
-from api.filtering.filtering import build_tag_query_dict_tuple
-from api.host_query_xjoin import owner_id_filter
+from api.filtering.filtering import query_filters
 from app import Permission
 from app.auth import get_current_identity
-from app.auth.identity import IdentityType
 from app.config import Config
 from app.environment import RuntimeEnvironment
-from app.exceptions import ValidationException
 from app.instrumentation import log_get_operating_system_failed
 from app.instrumentation import log_get_operating_system_succeeded
 from app.instrumentation import log_get_sap_sids_failed
@@ -133,24 +128,8 @@ def get_sap_system(tags=None, page=None, per_page=None, staleness=None, register
         "limit": limit,
         "offset": offset,
     }
-    hostfilter_and_variables = ()
 
-    if tags:
-        hostfilter_and_variables = ({"OR": build_tag_query_dict_tuple(tags)},)
-
-    if registered_with:
-        hostfilter_and_variables += build_registered_with_filter(registered_with)
-
-    if filter:
-        for key in filter:
-            if key == "system_profile":
-                hostfilter_and_variables += build_system_profile_filter(filter["system_profile"])
-            else:
-                raise ValidationException("filter key is invalid")
-
-    current_identity = get_current_identity()
-    if current_identity.identity_type == IdentityType.SYSTEM:
-        hostfilter_and_variables += owner_id_filter()
+    hostfilter_and_variables = query_filters(tags=tags, registered_with=registered_with, filter=filter)
 
     if hostfilter_and_variables != ():
         variables["hostFilter"]["AND"] = hostfilter_and_variables
@@ -182,30 +161,13 @@ def get_sap_sids(search=None, tags=None, page=None, per_page=None, staleness=Non
         "offset": offset,
     }
 
-    hostfilter_and_variables = ()
-
-    if tags:
-        hostfilter_and_variables = ({"OR": build_tag_query_dict_tuple(tags)},)
-
-    if registered_with:
-        hostfilter_and_variables += build_registered_with_filter(registered_with)
-
     if search:
         variables["filter"] = {
             # Escaped so that the string literals are not interpreted as regex
             "search": {"regex": f".*{custom_escape(search)}.*"}
         }
 
-    if filter:
-        for key in filter:
-            if key == "system_profile":
-                hostfilter_and_variables += build_system_profile_filter(filter["system_profile"])
-            else:
-                raise ValidationException("filter key is invalid")
-
-    current_identity = get_current_identity()
-    if current_identity.identity_type == IdentityType.SYSTEM:
-        hostfilter_and_variables += owner_id_filter()
+    hostfilter_and_variables = query_filters(tags=tags, registered_with=registered_with, filter=filter)
 
     if hostfilter_and_variables != ():
         variables["hostFilter"]["AND"] = hostfilter_and_variables
@@ -243,24 +205,8 @@ def get_operating_system(
         "limit": limit,
         "offset": offset,
     }
-    hostfilter_and_variables = ()
 
-    if tags:
-        hostfilter_and_variables = ({"OR": build_tag_query_dict_tuple(tags)},)
-
-    if registered_with:
-        variables["hostFilter"]["NOT"] = {"insights_id": {"eq": None}}
-
-    if filter:
-        for key in filter:
-            if key == "system_profile":
-                hostfilter_and_variables += build_system_profile_filter(filter["system_profile"])
-            else:
-                raise ValidationException("filter key is invalid")
-
-    current_identity = get_current_identity()
-    if current_identity.identity_type == IdentityType.SYSTEM:
-        hostfilter_and_variables += owner_id_filter()
+    hostfilter_and_variables = query_filters(tags=tags, registered_with=registered_with, filter=filter)
 
     if hostfilter_and_variables != ():
         variables["hostFilter"]["AND"] = hostfilter_and_variables
